clean up structure, set up lab and move classes into their own files
