+++ 2016-05/cort/lib/rank.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = function(name, rank, subRank, criteria, details) {
+    this.name = name;
+    this.rank = rank;
+    this.subRank = subRank;
+    this.criteria = criteria;
+    this.details = details;
+};
